Parse one item in 2 languages in one progress
What we want is an item in English and the Chinese translation which
could parse in the same way. Two items would go into ItemPineline
separately if we start request with two website, because only sigle
item can be accepted, it difficult to combine two items together,
especially parsing a ton of urls later. So we try to we start request
the English website and then get the Chinese one and combine two
languges websites' data together into one item, actually what we need
in Chinese website is only the transtion text - title, desc and details.
There is only one rule in both English and Chinese website. `rule`
here is the hostname which is the key in `website_rules`.